Reorganize APIs related to GraceDB uploads
