+++ pycbc/types/frequencyseries.py
@@ -398,7 +398,7 @@ class FrequencySeries(Array):
                                         self.numpy().imag)).T
             _numpy.savetxt(path, output)
         elif ext == '.xml' or path.endswith('.xml.gz'):
-            from pycbc.io.live import make_psd_xmldoc
+            from pycbc.io.ligolw import make_psd_xmldoc
             from ligo.lw import utils
 
             if self.kind != 'real':
